fix: DINOv3 float16 → bfloat16으로 변경 (attention NaN 해결)
float16은 exponent 범위가 좁아 eager attention의 softmax(exp() 연산)에서
overflow → NaN이 발생. bfloat16은 float32와 동일한 exponent 범위를 가져
NaN 없이 안정적으로 attention 계산 가능.

+++ models/dino.py
@@ -17,9 +17,9 @@ def load_dino(cfg: SimpleNamespace) -> tuple:
     processor = AutoImageProcessor.from_pretrained(cfg.model.dino_name)
     model = AutoModel.from_pretrained(
         cfg.model.dino_name,
-        torch_dtype=torch.float16,
+        torch_dtype=torch.bfloat16,    # bfloat16: same exponent range as float32 → no NaN in softmax
         device_map="auto",
-        attn_implementation="eager",
+        attn_implementation="eager",   # sdpa does not support output_attentions=True
     )
     model.eval()
     return model, processor
